Convert theshield JSON to CSV and append the result to the stripped.csv file.

+++ HostFilesUpdate.sh
@@ -113,7 +113,7 @@ fi
 ## Get DMRIds from two different sources, then merge them
 curl --fail -o /tmp/DMRIds_1.dat -s http://www.pistar.uk/downloads/DMRIds.dat
 curl --fail -o /tmp/DMRIds_2.dat -s http://registry.dstar.su/dmr/DMRIds2.php
-curl --fail -s "http://theshield.site/local_subscriber_ids.json" | python3 -c "exec('import sys, json\nresults = json.load(sys.stdin)[\'results\']\nfor entry in results:\n\ttry:\n\t\tprint(\'{}\\t{}\\t{}\'.format(entry[\'id\'], entry[\'callsign\'], entry[\'fname\'].encode(\'utf-8\', \'ignore\').decode()))\n\texcept:\n\t\tpass\n')" > /tmp/DMRIds_3.dat
+curl --fail -s "http://theshield.site/local_subscriber_ids.json" | tee >(python3 -c "exec('import sys, json\nresults = json.load(sys.stdin)[\'results\']\nfor entry in results:\n\ttry:\n\t\tprint(\'{}\\t{}\\t{}\'.format(entry[\'id\'], entry[\'callsign\'], entry[\'fname\'].encode(\'utf-8\', \'ignore\').decode()))\n\texcept:\n\t\tpass\n')" > /tmp/DMRIds_3.dat) >(python3 -c "exec('import sys, json\nresults = json.load(sys.stdin)[\'results\']\nfor entry in results:\n\ttry:\n\t\tplaceHolder=\"\"\n\t\tfname=\"\"\n\t\tcity=\"\"\n\\t\tstate=\"\"\n\t\tcountry=\"\"\n\t\ttry:\n\t\t\tfname=entry[\'fname\'].encode(\'utf-8\', \'ignore\').decode()\n\t\t\tcity=entry[\'city\'].encode(\'utf-8\', \'ignore\').decode()\n\t\t\tstate=entry[\'state\'].encode(\'utf-8\', \'ignore\').decode()\n\t\t\tcountry=entry[\'country\'].encode(\'utf-8\', \'ignore\').decode()\n\t\texcept:\n\t\t\tpass\n\t\tprint(\'{},{},{},{},{},{},{},{}\'.format(entry[\'id\'], entry[\'callsign\'], fname, placeHolder, city, state, country, placeHolder))\n\texcept:\n\t\tpass')" > /tmp/stripped.csv) > /dev/null 2<&1
 cat /tmp/DMRIds_1.dat /tmp/DMRIds_2.dat /tmp/DMRIds_3.dat | grep -v ^# | awk '($1 > 9999) && ($1 < 10000000) { print $0 }' | sort -un -k1n -o ${DMRIDFILE}
 rm -f /tmp/DMRIds_1.dat /tmp/DMRIds_2.dat /tmp/DMRIds_3.dat
 
@@ -132,7 +132,11 @@ curl --fail -o ${TGLISTYSF} -s http://www.pistar.uk/downloads/TGList_YSF.txt
 
 # NextionDriver stuff
 curl --fail -k -o /usr/local/etc/groups.txt -s https://api.brandmeister.network/v1.0/groups/
-curl --fail -o /usr/local/etc/stripped.csv -s https://database.radioid.net/static/user.csv
+if [ ! -e /root/NO_STRIPPED_FILE ]; then
+	curl --fail -o /usr/local/etc/stripped.csv -s https://database.radioid.net/static/user.csv
+	cat /tmp/stripped.csv >> /usr/local/etc/stripped.csv
+fi
+rm -f /tmp/stripped.csv
 
 
 # If there is a DMR Over-ride file, add it's contents to DMR_Hosts.txt
